pyproject.toml
[project]
name = "qbot"
version = "0.1.0"
description = "QQ score statistics bot based on NoneBot2 + OneBot v11"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "nonebot2[fastapi]>=2.3.0",
  "nonebot-adapter-onebot>=2.4.0",
  "nonebot-plugin-apscheduler>=0.5.0",
  "aiosqlite>=0.20.0",
  "matplotlib>=3.8.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"
